backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-cors-headers>=4.0
psycopg2-binary>=2.9
Pillow>=10.0
python-dotenv>=1.0
dj-database-url>=2.1
whitenoise>=6.5
gunicorn>=21.2